requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.13.0
peft>=0.4.0
accelerate>=0.20.0
bitsandbytes>=0.40.0
pyyaml>=6.0
protobuf>=3.20.0
tiktoken>=0.5.0
sentencepiece>=0.1.99



